BETA FIX
rewrote code to fix NullPointerException
